+++ operator/pkg/apis/cassandra/v1beta1/cassandradatacenter_types_test.go
@@ -4,201 +4,12 @@
 package v1beta1
 
 import (
-	"os"
 	"testing"
 
-	"github.com/datastax/cass-operator/operator/pkg/images"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-func Test_makeImage(t *testing.T) {
-	type args struct {
-		serverType    string
-		serverImage   string
-		serverVersion string
-	}
-	tests := []struct {
-		name      string
-		args      args
-		want      string
-		errString string
-	}{
-		{
-			name: "test empty image",
-			args: args{
-				serverImage:   "",
-				serverType:    "dse",
-				serverVersion: "6.8.0",
-			},
-			want:      "datastax/dse-server:6.8.0",
-			errString: "",
-		},
-		{
-			name: "test empty image cassandra",
-			args: args{
-				serverImage:   "",
-				serverType:    "cassandra",
-				serverVersion: "3.11.7",
-			},
-			want:      "datastax/cassandra-mgmtapi-3_11_7:v0.1.13",
-			errString: "",
-		},
-		{
-			name: "test private repo server",
-			args: args{
-				serverImage:   "datastax.jfrog.io/secret-debug-image/dse-server:6.8.0-test123",
-				serverType:    "dse",
-				serverVersion: "6.8.0",
-			},
-			want:      "datastax.jfrog.io/secret-debug-image/dse-server:6.8.0-test123",
-			errString: "",
-		},
-		{
-			name: "test unknown dse version",
-			args: args{
-				serverImage:   "",
-				serverType:    "dse",
-				serverVersion: "6.7.0",
-			},
-			want:      "",
-			errString: "server 'dse' and version '6.7.0' do not work together",
-		},
-		{
-			name: "test unknown cassandra version",
-			args: args{
-				serverImage:   "",
-				serverType:    "cassandra",
-				serverVersion: "3.10.0",
-			},
-			want:      "",
-			errString: "server 'cassandra' and version '3.10.0' do not work together",
-		},
-		{
-			name: "test fallback",
-			args: args{
-				serverImage:   "",
-				serverType:    "dse",
-				serverVersion: "6.8.1234",
-			},
-			want:      "datastax/dse-server:6.8.1234",
-			errString: "",
-		},
-		{
-			name: "test cassandra fallback",
-			args: args{
-				serverImage:   "",
-				serverType:    "cassandra",
-				serverVersion: "3.11.1234",
-			},
-			want:      "datastax/cassandra-mgmtapi:3.11.1234",
-			errString: "",
-		},
-		{
-			name: "test 6.8.4",
-			args: args{
-				serverImage:   "",
-				serverType:    "dse",
-				serverVersion: "6.8.4",
-			},
-			want:      "datastax/dse-server:6.8.4",
-			errString: "",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := makeImage(tt.args.serverType, tt.args.serverVersion, tt.args.serverImage)
-			if got != tt.want {
-				t.Errorf("makeImage() = %v, want %v", got, tt.want)
-			}
-			if err == nil {
-				if tt.errString != "" {
-					t.Errorf("makeImage() err = %v, want %v", err, tt.errString)
-				}
-			} else {
-				if err.Error() != tt.errString {
-					t.Errorf("makeImage() err = %v, want %v", err, tt.errString)
-				}
-			}
-		})
-	}
-}
-
-func Test_makeUbiImage(t *testing.T) {
-	type args struct {
-		serverType    string
-		serverImage   string
-		serverVersion string
-	}
-	tests := []struct {
-		name      string
-		args      args
-		want      string
-		errString string
-	}{
-		{
-			name: "test fallback",
-			args: args{
-				serverImage:   "",
-				serverType:    "dse",
-				serverVersion: "6.8.1234",
-			},
-			want:      "datastax/dse-server:6.8.1234-ubi7",
-			errString: "",
-		},
-		{
-			name: "test cassandra fallback",
-			args: args{
-				serverImage:   "",
-				serverType:    "cassandra",
-				serverVersion: "4.0.1234",
-			},
-			want:      "datastax/cassandra-mgmtapi:4.0.1234-ubi7",
-			errString: "",
-		},
-		{
-			name: "test unknown dse version",
-			args: args{
-				serverImage:   "",
-				serverType:    "dse",
-				serverVersion: "6.7.0",
-			},
-			want:      "",
-			errString: "server 'dse' and version '6.7.0' do not work together",
-		},
-		{
-			name: "test unknown cassandra version",
-			args: args{
-				serverImage:   "",
-				serverType:    "cassandra",
-				serverVersion: "3.10.0",
-			},
-			want:      "",
-			errString: "server 'cassandra' and version '3.10.0' do not work together",
-		},
-	}
-	for _, tt := range tests {
-		os.Setenv(images.EnvBaseImageOS, "example")
-
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := makeImage(tt.args.serverType, tt.args.serverVersion, tt.args.serverImage)
-			if got != tt.want {
-				t.Errorf("makeImage() = %v, want %v", got, tt.want)
-			}
-			if err == nil {
-				if tt.errString != "" {
-					t.Errorf("makeImage() err = %v, want %v", err, tt.errString)
-				}
-			} else {
-				if err.Error() != tt.errString {
-					t.Errorf("makeImage() err = %v, want %v", err, tt.errString)
-				}
-			}
-		})
-		os.Unsetenv(images.EnvBaseImageOS)
-	}
-}
 
 func TestCassandraDatacenter_GetServerImage(t *testing.T) {
 	type fields struct {
@@ -244,20 +55,10 @@ func TestCassandraDatacenter_GetServerImage(t *testing.T) {
 				Spec:       tt.fields.Spec,
 				Status:     tt.fields.Status,
 			}
-			got, err := dc.GetServerImage()
+			got := dc.GetServerImage()
 			if got != tt.want {
 				t.Errorf("CassandraDatacenter.GetServerImage() = %v, want %v", got, tt.want)
 			}
-			if err == nil {
-				if tt.errString != "" {
-					t.Errorf("CassandraDatacenter.GetServerImage() err = %v, want %v", err, tt.errString)
-				}
-			} else {
-				if err.Error() != tt.errString {
-					t.Errorf("CassandraDatacenter.GetServerImage() err = %v, want %v", err, tt.errString)
-				}
-			}
-
 		})
 	}
 }
